Improving Net interface

+++ src/main/java/org/zero/utils/PrintPicture.java
@@ -2,9 +2,9 @@ package org.zero.utils;
 
 import java.util.function.Consumer;
 
-public class PrintPicture implements Consumer<int[]> {
+public class PrintPicture implements Consumer<double[]> {
     @Override
-    public void accept(int[] images) {
+    public void accept(double[] images) {
         for (int i = 0; i < 28; i++) {
             for (int j = 0; j < 28; j++) {
                 System.out.print(images[i * 28 + j] > 0 ? "%" : " ");
